+++ src/main/java/ru/r2cloud/jradio/blocks/LowPassFilter.java
@@ -39,7 +39,7 @@ public class LowPassFilter implements FloatInput {
 			samples = null;
 		}
 		context = new Context(source.getContext());
-		context.setSampleRate((long) context.getSampleRate() / decimation);
+		context.setSampleRate(context.getSampleRate() / decimation);
 		context.setTotalSamples(context.getTotalSamples() / decimation);
 	}
